mv_dma: setup address decoding windows
This is a bit tricky as IDMA on one hand wants to find a decoding window
for source as well as target address, while TDMA on the other hand
complains if source and target are both found in decoding windows.
Therefore setup decoding windows for DRAM access and an additional one
for SRAM access on IDMA only.